Added gcc/clang compatibility layer for coroutines
This commit adds a file called stdlib_coroutine.hpp, which includes
either <coroutine> or <experimental/coroutine> based on which standard
library is used. If necessary for that compiler, it also includes
std::experimental::<coroutine_stuff> in std, or vice versa.

This enables instances of `std::experimental` to be replaced with
`std` in the code base. So rather than using
std::experimental::coroutine_handle, we can use std::coroutine_handle.

Signed-off-by: Alecto Perez <perez.cs@pm.me>