Removed fmt module
Having *fmt* added as a git module isn't necessary for the project. If
no system installation for fmt can be found, CMake will automatically
download one as part of the build process. The master branch also
removed the fmt module, so I thought it appropriate to remove it here
as well.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>

1 commit.