Added .clang-format
This commit adds a .clang-format to the project, and formats code files
according to the specification outlined in the .clang-format. This
helps to make the code more readable, and to make things more
consistent.

NB: In VSCode, any code file can be formatted with Ctrl+Shift+I, and
it'll automatically be formatted according to the clang-format.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>